requirements.txt
# Core
Django==3.2.23  # LTS version that works with djongo
python-dotenv==1.0.0

# Database
pymongo==3.13.0
djongo==1.3.6

# REST Framework
djangorestframework==3.14.0
django-cors-headers==3.14.0

# Authentication
djangorestframework-simplejwt==5.3.0

# Utilities
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
python-dateutil==2.8.2
python-magic==0.4.27

# Development
django-debug-toolbar==4.1.0
pytest==7.4.2
pytest-django==4.7.0

# Production
gunicorn==21.2.0
whitenoise==6.5.0

# Celery (if needed)
celery==5.3.4
gunicorn
brotli
requests[brotli]
# Other
django-environ==0.11.2
Pillow==10.0.1
python-slugify==8.0.1
openai
aiohttp
langchain
langchain_community
scikit-learn
pymongo[srv]

